fixed integer rounding causing jittering, alpha bug, and tile spacing

+++ src/util.rs
@@ -3,8 +3,10 @@ use std::collections::HashMap;
 use core::Camera;
 
 pub fn world_to_pixel(w_x: f32, w_y: f32, camera: &Camera) -> (u32, u32) {
-    let p_x = (w_x - camera.x) as i32 * camera.tile_pix as i32 + (camera.canvas_width/2) as i32;
-    let p_y = (w_y - camera.y) as i32 * camera.tile_pix as i32 + (camera.canvas_height/2) as i32;
+    //let p_x = (w_x - camera.x) as i32 * camera.tile_pix as i32 + (camera.canvas_width/2) as i32;
+    //let p_y = (w_y - camera.y) as i32 * camera.tile_pix as i32 + (camera.canvas_height/2) as i32;
+    let p_x = (w_x - camera.x) * camera.tile_pix as f32 + (camera.canvas_width/2) as f32;
+    let p_y = (w_y - camera.y) * camera.tile_pix as f32 + (camera.canvas_height/2) as f32;
     (p_x as u32, p_y as u32)
 }
